requirements.txt
flask>=2.0.0
pandas>=1.3.0
sentence-transformers>=2.1.0
pinecone-client>=2.0.0
pytesseract>=0.3.8
pillow>=8.3.0
requests>=2.25.1
beautifulsoup4>=4.9.3
tensorflow
